requirements.txt
typer==0.4.1
inquirer==2.9.1
PyYAML==6.0